requirements.txt
requests==2.28.1
pydantic==1.10.2
pyTelegramBotAPI==4.7.0
python-dotenv==0.21.0
py_markdown_table==0.3.3
python-Levenshtein==0.20.5
unidecode==1.3.6